+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.19.0] - 2025-09-03
+
 - ✨(front) add modal update mailboxes #954
 
 ### Added
@@ -435,7 +437,8 @@ and this project adheres to
 - ✨(domains) create and manage domains on admin + API
 - ✨(domains) mailbox creation + link to email provisioning API
 
-[unreleased]: https://github.com/suitenumerique/people/compare/v1.18.2...main
+[unreleased]: https://github.com/suitenumerique/people/compare/v1.19.0...main
+[1.19.0]: https://github.com/suitenumerique/people/releases/v1.19.0
 [1.18.2]: https://github.com/suitenumerique/people/releases/v1.18.2
 [1.18.1]: https://github.com/suitenumerique/people/releases/v1.18.1
 [1.18.0]: https://github.com/suitenumerique/people/releases/v1.18.0